pyproject.toml
[project]
name = "gh-audit"
version = "0.1.3"
readme = "README.md"
authors = [{name = "Joshua Peek"}]
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.1",
    "pygithub>=2.0.0",
    "pyyaml>=6.0.1",
]
classifiers = [
    "License :: OSI Approved :: MIT License",
]

[dependency-groups]
dev = [
    "mypy>=1.0.0,<2.0",
    "ruff>=0.4.0",
    "types-PyYAML>=6.0.1",
]

[project.scripts]
gh-audit = "gh_audit.cli:main"

[tool.ruff.lint]
extend-select = ["I", "UP"]

[tool.mypy]
strict = true

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
